requirements.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==2.3.1
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
coverage==7.6.0
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-silk==5.1.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==26.0.0
filelock==3.15.4
gprof2dot==2024.6.6
identify==2.6.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy==1.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240712
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
